app/docker/requirements.txt
boto3==1.37.33
botocore==1.37.33
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
jmespath==1.0.1
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
urllib3==2.4.0